refactor: review suggestions
