+++ packages/gatsby-theme-docs/src/components/index.js
@@ -5,7 +5,7 @@ export { default as ContentPagination } from './content-pagination';
 export { default as GlobalNavigationLink } from './global-navigation-link';
 export { default as ErrorBoundary } from './error-boundary';
 export { default as Link, ExternalSiteLink } from './link';
-export { default as MultiLanguageCodeExamples } from './multi-language-code-examples';
+export { default as MultiCodeExample } from './multi-code-example';
 export { default as Overlay } from './overlay';
 export { default as SEO } from './seo';
 export { default as SearchDialog } from './search-dialog';
